chore: Add iproxy alias for setting up a proxy server
The .aliases file was updated to include a new alias, iproxy, which sets up a proxy server. This alias generates a random login and password, displays them, and sets up the HTTPS_PROXY environment variable with the generated credentials. It also starts a Docker container running the 3proxy image. This change improves the development workflow by providing an easy way to set up a proxy server for testing purposes.