chore: Update .profile with new environment variables
Add new environment variables BUN_CONFIG_NO_CLEAR_TERMINAL_ON_RELOAD and DO_NOT_TRACK to the .profile file. These variables are set to "1" and will be used for specific configurations related to the BUN package manager and tracking preferences. This change ensures that the necessary environment variables are set for the desired behavior.