chore: Update .zshrc aliases and add kubecolor plugin
Refactor the aliases for the ls and cat commands in the .zshrc file to use conditional statements based on the availability of the 'eza' and 'bat' commands, respectively. Additionally, add the 'kubecolor' plugin to the Zinit configuration. These changes improve consistency, compatibility, and enhance the readability of Kubernetes command output.