chore: Refactor .stuff/README.md to update Ansible playbook task
Update the .stuff/README.md file to remove the TODO comment and modify the task description. The changes involve replacing the phrase "create an Ansible playbook" with "automatise this + get the client.ovpn" to accurately reflect the purpose of the task. This improves the clarity and understanding of the playbook's objective.