chore: Update .stuff/README.md to reflect changes in task description
Remove the outdated "TODO" comment and update the task description in the .stuff/README.md file. The phrase "create an Ansible playbook" has been replaced with "automatise this + get the client.ovpn" to accurately reflect the purpose of the task. This improves the clarity and understanding of the playbook's objective.